Moved PHThread to a new file
